Hide fluid-specific constructs in non-fluid.
As a transitional step between fluid and non-fluid, usrs may try fluid, get
confused, and then switch back to non-fluid. However, if they choose a
fluid-specific construct while doing so (including by accident), their code will
break in non-fluid. This unwraps the fluid constructs, so that if you make a
change in non-fluid, it will overwrite the old fluid construct.