This is the tihng that shows up in the alerts channel
Right now, the "revision" field is the docker image id
(gcr.io/balmy-ground-195100/fe17440515d9), which seems not terribly
useful? I'm replacing it with the git sha, which is more useful.

Prevoiusly I'd put that in the comment, along with the circleci link
(https://circleci.com/gh/darklang/dark/27107), separated by a
semi-colon, but it seems that rollbar just stops at the semi-colon.

(Since we have the circleci link, if you want a link to the docker
image, you can get it from the circle build scripts.)