Better error for stack overflow in handler analysis
A few things going on here:
- we see this error when an analysisResult's function_result contains an
overly large DList
- It's possible other errors could cause this; see comment in code for
why I didn't scope the catch more tightly
- We want to present an error that is more human-friendly than just
  "Stack overflow" (or SyntaxError: Invalid regular expression: /maximum
  call stack/: Maximum call stack size exceeded", which also
  occurred)
- In the error, I am trying to convey:
  - This is a known bug (stack overflow while deserializng overly large
    function results)
  - Which handler it is in (_should_ be obvious from "you just clicked
    it", but our errors could use some work here, in the spirit of Elm/Rust
  - Your code should work fine in bwd!

This makes for a somewhat long error message. Open question whether we
prefer long verbose errors or do a rust-style "here's the short error,
here's a code (or link, since we're in-browser) for more details about
what this means". Not gonna address that now, but call it product
discussion fodder for later.

https://trello.com/c/1HakZf5x/1570-max-call-stack-size-errors