Integration test showing how to trigger a handler
The darklang code being tested here is redundant with the backend test
`Crypto::sha256hmac works for aws`, _but_ its value as an integration
test is that it demonstrates how to run a REPL in testcafe. (I suspect
this would also work for HTTP handlers.)

I'm checking it in because it was really hard to get working; I still
don't understand _why_ I had to call
`click(Selector('div.handler-trigger'))` twice, but I think it's worth
documenting so we can have integration tests in the future that do this
sort of thing.

https://trello.com/c/r2Ee2hLD/1941-document-how-to-click-the-handler-trigger-button-in-integration-tests-for-future-reference