Add rust-integration-tests CI step
The queue-scheduler tests rely on the fact that the testing database
exists and has the proper schema already. To accomplish this, we add a
new step in parallel to the existing integration-tests, which relies on
the backend, client, and queue-scheduler builds, pulling in their built
artifacts via cache and workspace.