Integration test for FE toBytes implementation
Prior to this PR/commit, I reverted the "Partial hoist of lookup" commit
(see 76a9508f48e3df1f48361324746325a815adc590 for the revert), which
broke client-side toBytes/toString roundtripping.

I noticed it manually while prepping for retro and was able to revert
the offending commit pre-merge, but this integration test should defend
against that in the future, as well as if we try later to clean up that
js code.

Possibly it could be a FE-only test, I'm not sure, but this works for
now.

Followup to https://trello.com/c/CkhSc3Ox/1859-unblock-eagon-and-other-s3-users-by-exposing-functions-for-aws-signing