Fix race cond between RPC req + stroller messages in Function editing
This is the thing I'd previously fixed in Handlers (#1299, #1331,
https://trello.com/c/79pqHMfl/1495-fix-flicker-from-rpc-race), there was a bug
causing it not to be applied to TLs that were functions.

NOTE: the bug here is "you applied the new handler(s), and then did the
check for getting a tlid out of a model [intending to get the old func
from the old model], but you already updated the model. Boo.", so an
_order of mutation_ problem; it is fixed by:
- preserving the oldM before mutating it, so we can later retrieve the
old func, AND
- using the same function we use for this (bringBackCurrentTL) in the
SetToplevels case, instead of having the ~same code twice

Test: if I create a function with `let x = "aaaaaaaaaaaaaaaaaaaaa|"`, and
then just hold down delete within the string, my cursor outraces the end
of the string, and I end up with, say, `let x = "|aaa"`.  Likewise, if I
start with `let x = "|"` and hold down 'a', cursor outraces string, and
i end up with `let x = "aaaaaaaaaaaaa"\naaaaaaaa` (that is, the body
contains an unbound variable `aaaaaaa`).

After this commit, behavior is as expected.

No automated test.

https://trello.com/c/mHCLZpXE/1973-cant-type-fast-in-functions