Default worker pause button to "run" state.
This has the effect of making newly created workers correctly show up
with the pause button. The problem here is that the `workerStats` state
does not include the newly created worker, so it was getting rendered
without the button.

Note that this is only correct so long as the server creates workers in
the "run" state and if the server starts creating the handler paused,
this will need to change.