Make docker reap orphan processes in scripts/builder
tl;dr: pid 1 is responsible for reaping orphans, docker breaks this if
you have a multiprocess container. Which is why, in our local
scripts/builder container, we end up with lots of <defunct> processes.

Not a functional problem, but a minor annoyance if one is running ps to
check up on one's dev env. Solved by using docker's --init flag. Only
affects the local env, because this is in scripts/builder.

(See https://tech.fpcomplete.com/rust/pid1 for more about pid 1 and
various solutions to this problem; it's how I discovered docker --init,
which is doc'd at
https://docs.docker.com/engine/reference/run/#specify-an-init-process)

https://trello.com/c/Uh2vEqMh/2076-use-docker-init-to-avoid-orphan-processes-in-dev-container