Better document how to add a custom domain for a user, and why we don't support apex domains
Also, script the code changes! The process still requires a manual
deploy, but that is clearly documented, and there are fewer steps than
before.

Finally, the script also tells you how many certs we're using out of the
max 15 Google allows, before we'll need to do infra work on this.

https://trello.com/c/qIGFZY6L/2068-domain-for-alex-cremner