404s on /api/:canvas/get_trace_data, but should not rollbar
As of 213f32bfe4 (2019-11-23), Paul made all (or many of?) our Js.logs
reportError, with the intent that errors should be rollbar-visible, not
hidden from us.

And we treat 404s in FetchTraceData as an error, because that indicates
"we haven't saved the handler yet" (see comment above on
workers/Fetcher.ml:44), and returning TraceFetchMissing will cause the
worker to retry. This is good!

However, it is _not_ an error in the sense that it is an expected event,
and it is sufficiently high-frequency that it floods rollbar.

To fix, we just use remove the reportError line! We can still, if we
want, see 404s on this endpoint using honeycomb (WHERE request_pathshape
= /api/:canvas/get_trace_data GROUP BY status).

https://trello.com/c/wM9ejv4M/2088-get-rid-of-flood-of-tracefetchmissing-errors-to-rollbar