Fix ac ordering
"The problem is that we order them alphabetically, but we include the version suffix"

(That is, `dbget` matched "DB::getAll_v1" ahead of "DB::get_v1",  because A < _.)

We fix this by splitting on '_v' and then sorting by the resulting list.
So now we're comparing ["DB::get"; "1"] against ["DB::getAll"; "1"]. (We
also to_lower the string, since we still want case-insensitive sorting
and matching.)

We could call List::head and just compare "DB::get" with "DB::getAll",
but then we'd lose any sorting of "DB::get_v1" vs "DB::get_v2".

https://trello.com/c/UxvKtzWw/2083-dbget-should-be-highlighted-instead-of-dbgetall-when-its-the-exact-match