Make nearly all clicking work properly in Fluid.
This abstracts the midClick behaviour into two Modification: FluidStartClick and
EndClick.

It also solves a bug that we were seeing in applying FluidMouseUp and
FluidMouseDown, which is that they reused the old model instead of the latest
one. I added an Apply modification, to make sure we use the latest model for
this.

This allowed me to roll the FluidMouseUp changes into Fluid.ml