https://trello.com/c/6YvEjT9x/2040-most-recent-db-entry-and-connected-workers-handlers-appear-intermittently-in-fluid
add border to make it (maybe) more clear that the top is draggable and bottom is not

Turn TLDragRegionMouseUp's that don't move the TL into clicks

fix integration test

Not sure why this is necessary, but w/o it, I consistently get
"specified selector does not match any node in the DOM tree" on the
second read of ".selected .live-value"

Make the draggable area larger

Paul noticed (and we repro'd) that there's a large area around the edges
of a DB handler  that aren't draggable. Explanation for why this is, and
how we fixed it, is in a comment in the commit.