strip :: when figuring out autocomplete
it's common to type something like `dbget` when looking for `DB::get_v1`
but the prior implementation took the separator into account when
looking for matches, meaning that `DB::generateKey` was actually the
first match for `dbget`.

Instead, strip the :: separator from both function names and the query
when comparing.