Different solution for wildcard domains for integration tests
Each integration tests uses a different canvas, each of which have different
URLs. The general format is
http://test-sometest.builtwithdark.localhost:8000. We previously ran dnsmasq to
allow these wildcard domains which are not supported in /etc/hosts.

Unfortunately, the optimizations we were making to our CircleCI config did not
allow dnsmasq to run (it needed the NET_ADMIN capability, which CircleCI does
not provide in Docker mode). So several of the integration tests started to
fail.

A solution is to use lvh.me, which always resolves to 127.0.0.1, as do all of
its subdomains. Given that this is only used for integration tests, this doesn't
seem to be a security risk.

This fixes the integration tests.