Allow dark user's uid (and group's gid) to be set at build time
By exposing uid/gid as build-args, we can set these values to match the
host user's uid/gid - allowing for dark-owned files in-container to be
owned by the host user on the host fs.

We didn't need this in OS X because Docker for Mac handles it for you,
and we avoided it on Linux for a while because often the first non-root
user has uid 1000. But that's not always the case, and when it's not,
you get files owned by 1000:1000, which the host can't deal with. (And
so you need to chown files on the host and the guest.)