Convert DBlock to `DBlock of string list * expr`
The string list are the names of the lambda's parameters. The expr is the body
of the lambda.

The lambda body is now evaluated by whatever is trying to execute the
lambda. The exec_state now has an exec function and the latest symtable to help
with that.

Does serialization on the frontend to support execute_function. Probably needs
work to get serialization right on the backend.

Stdlib still needs to be converted.

The sql_compiler has been fully shifted over, though doesn't work yet.