actual string escaping for n, r, t, '"'
We don't special-case \; a single \ stroke will result in a valid
string, with a valid backslash.

As such, we don't actually move between EString and EPartial; strings
are valid regardless.

NOTE: the live value of a string will always show the 'correct' thing -
that is, \n will be a newline, not the string '\n'.