Distinguish omnibox pos
Distinguis when omnibox is opened by user click (when user has intentionally opened it at a specific spot on the canvas)

vs. when user hit ENTER/Cmd+k (should be centered)

When user opens omnibox, with intended location we should continue to place the new toplevel near where they open the omnibox

When user does not have a location in mid, omnibox should open in the "center top" of the screen, and new toplevels created will be place using the same logic as (+) button from sidebar, placing it within viewport at a random location.