events needs partial index where status <> done
We don't need the events index to cover events with status = 'done'; not
indexing those keeps the index small enough to fit in cache.

This was responsible for the worker queues outage of 2020-01-09.

Solution: `create index concurrently idx_events_for_dequeue on events
(status, id) where status <> 'done'` in prod, manually; migration to
create index if not exists. (Same index, but not concurrently because
our migration script doesn't handle that well.)

https://trello.com/c/u5UusTxv/2212-partial-index-on-events-queue