Handle 'thread killed' error
If a thread is killed in our two-thread workers (qw, cron checker), it
doesn't have a chance to set the shutdown ref, and so the other thread
stays running. This means an unhandled exn in the actual worker (qw,
cron_checker) doesn't terminate the pod, because the healthcheck thread
stays up ... so we have a pod not processing work but also not being
replaced with a new one.

Orig log (there've been other similar ones):
```
Thread 2 killed on uncaught exception Libexecution__Exception.DarkException(_)
[Stack trace]
```

Three cases where we want to exit:
- healthcheck worker is instructed to die (/pkill), it sets the shutdown
  ref, cron_checker loop terminates
- heathcheck worker dies/is killed (unhandled exn), kubernetes will kill
  the pod when it fails healthcheck
- cron_checker thread dies; it's the main loop, the process exits