When copy/pasting, just keep the entire oldExpr of a partial.
We had a problem where the contents of a partial are lost if you highlight
it. The problem is that the old name is hidden by the new name, so it is hard to
reconstruct it properly.

In addition, what should happen if you don't select the arguments? Losing the
arguments actually breaks the partial, so it's probably best to keep them.