DB::get_v2, when it fails, should not log an exn to honeycomb
DB::get_v2 is not expected to always work - it returns an option! Our
implementation assumes otherwise, and treats "no record for this key" as
an exception.

In theory, we handled this exception, but this is not great for two
reasons:
- DB exns get logged and then re-raised as DarkExceptions (possibly to
get the behavior we saw here? I'm gonna briefly look to see if there are
others we should prune), so a non-exceptional behavior got logged as an
exception
- exns are expensive

Additionally, I added Dval.dopt_of_option; I want User_db.get_option to
return a `dval option` (ocaml's option type) because if it returns a
DOption, ocaml's type is dval, which is too broad.

https://trello.com/c/ApbpylVp/2221-dbgetv2-and-v1-should-not-raise-log-a-dark-exception-its-return-type-is-doption