Lower CPU reqs for bwd
We're about to have more pods running in our cluster (easier to spin up
two replicasets of scale=72 than to figure out where traffic actually
is; we can move those #s later if we choose). Paul reports that
we are way under-using CPU (per stackdriver), so the thing that is a
bottleneck for us right now is CPU requests, not actual usage. (Per:
`kubectl describe nodes | grep Resource -A 3`, and eyeballing the
results.)

This halves the CPU requests/limits for bwd, making room to do that.

It does not touch memory reqs/limits, because we're not bumping up
against that.

Deploy note: I'm going to deploy this manually and watch for a bit
before I merge.