When we have the traceID but no trace for it, request the trace.
Before this, when we selected the traceID, if there was no trace for it, we
would drop the info that we knew the traceID, and not request an analysis.

This seems to have been wrong but stable behaviour that must have gotten
dislodged by a recent change. Perhaps the ordering of things changed, or it took
longer for a worker to run. It's pretty clear it should have been like this all
along though.