JSON::parse should return a result, not error
This came out of bogdan-dynad, which runs a 1/minute CRON that parses a
string value from a datastore that is not valid json. Since JSON::parse
raises an exception, it bubbles up to the queue worker, which logs and
rollbars.

Related PRs, with logging that got us to this diagnosis:
https://github.com/darklang/dark/pull/1868 , https://github.com/darklang/dark/pull/1866

I deliberately chose to make this change to parse_v0 instead of a new
function version:
- auto-unwrapping from the error rail means that existing code should
  continue to work as-is (valid json needs no extra handling; invalid
  json now error-rails, which makes the error more visible to the user but
  doesn't demand that they handle it)
- existing code raising leads to exceptions in our code, not the user's,
  which is inappropriate (improper json is their problem, not ours)
  - If we added parse_v1 and deprecated v0, we'd have to go upgrade
    users to fix the constant rollbar problem

https://trello.com/c/KeihVmff/2252-jsonparse-on-invalid-json-should-return-a-reserrornot-raise