Add menu state in cache
Before when menu state was in handlerProps contributed to the TL cache key.
To generalize menus, I took it out of handlerProps and put it in its own TLIDDict.
We want to add that state back in so we will render the open/close menu state properly for all TLs.

Also change open listener to back to onClick, because it does seem alot of our analysis (such as if DB is used in code, depends if the TL is selected).