To impl the auth /callback in Darklang, we need to be able to put a session in the DB
We can't use Auth.Session.new_for_username like webserver.ml, because
that fn is lwt, and our interpreter is not.

I considered just exposing `INSERT INTO session`, since the inputs to
the SQL query are just random strings of a given length ... but this
seemed like a cleaner boundary to draw. (And avoids having to wonder
whether Darklang's String::random is cryptographically secure, and
whether that matters.)

Also considered: we could make this a userdb belonging to the ops-login
canvas, and rewrite Auth.Session to query the user_data table
appropriately. Call that a possible future migration (if we don't switch
to using JWTs first).

https://trello.com/c/htk6fNlz/2309-darkinternalnewsessionforusername