Don't pctencode the cookie name or value in Http::setCookie
Session.to_cookie_hdrs doesn't (as tested by generating a bunch of
sessions locally, and seeing that the cookie header was unescaped:
"set-cookie: __session=DLz/MRcJfGehC5DZpnmWETXhWjNP/p+iSPjdH16T;
Max-Age=604800; domain=localhost; path=/; secure; httponly"

instead of escaped (we'd expect, if it were escaped, to have seen the
'/' (two) in the above be '%2F').

This was discovered while testing out ops-login.darklang.com, which uses
Http::setCookie (darklang) instead of Session.to_cookie_hdrs (ocaml).
Specifically, when we generated a session whose key contained url-unsafe
chars, the DB would store the unescaped value but we'd tell the browser to
set the escaped value as its cookie, and so auth would break.

RFC DISCUSSION:
For this, let's not worry about the allowed charset of a cookie-name; it
is not randomly generated, and we can
expect the user to give Http::setCookie an appropriate cookie-name
(like, in our case, `__session`).

So let's discuss cookie values, which actually tickled this bug!
Per
https://tools.ietf.org/html/draft-ietf-httpbis-rfc6265bis-02#section-4.2:
```
   cookie-value      = *cookie-octet / ( DQUOTE *cookie-octet DQUOTE )
   cookie-octet      = %x21 / %x23-2B / %x2D-3A / %x3C-5B / %x5D-7E
                         ; US-ASCII characters excluding CTLs,
                         ; whitespace DQUOTE, comma, semicolon,
                         ; and backslash
```

I think it is, again, reasonable to expect the user to give
Http::setCookie an appropriate cookie-value.

And in fact, the two places where we use Http::setCookie in ops-login do
this:
- DarkInternal::newSessionForUsername: uses ocaml-session, let's assume
  that DTRT; we saw this bug because ocaml-session gave us a
  forwardslash, which is an acceptable cookie-octet, but not url-safe
  (and so we pctencoded it)
- String::random_v2 (for the __auth cookie): this uses
  Util.random_string, which selects from chars [a-zA-Z0-9], all in the
  safe set

CURRENT USE:
According to get_functions_and_tlids.exe, as of last prodclone, there
were 20 uses in prod of setCookie.

Of these (and ignoring the ones that can't load anyway b/c DuplicateDB
error), the cookie value is:
- a base64-encoded String::random
- empty string ("unset cookie")
- a jwt (one instance, reify-modification-migrator)
- a string literal (one instance, in a REPL, the value "hi")

Making Http::setCookie check its value's validity and
return a result is possible, but out of scope for this change.

https://trello.com/c/fzzt0Dkc/2318-httpsetcookie-bug-should-not-pctencode-cookie-name-and-value