Logout link now redirects to login.darklang.com/logout
That page handles the auth0 logout if an access_token is found (in
ops-login's AccessToken datastore), and also clears the session from
postgres. When done, we redirect to darklang.com/login

As such, it is backwards compatible with the existing, pre-auth0,
sessions, and so we can deploy it now, no need to wait for the auth0
rollout.

https://trello.com/c/rOy3ESrN/2347-new-logout-link