List::sortByCmp f
We have List::sortBy, which takes an f 'a -> 'b and sorts by the output.

This new thing takes f 'a 'a -> (0,1,-1) and sorts by the comparator, as
is necessary when any two inputs can be ordered relative to each other,
but you can't come up with a single orderable value.

I almost have an in-dark implementation of semver sorting, but it
depends on this - I'll put screenshots in the PR since I think that'll
help clarify what's going on here.

https://trello.com/c/pqVeddlx/2365-listsortbycomparator