Fix field AC triggering repeatedly
This caused a recoverable assertion failure. After accepting a field
autocomplete, the `query` was not reset, so acMaybeCommit always fired,
resulting in the E.replace in acEnter being called. However, the token
info partial that was in ac.query was gone by then (it had been turned
into a non-partial the first time the AC was accepted), so it repeatedly
threw an error.