integration tests
[x] happy path: can upload dark/stdlib/Test::uuid_v[something]
[x] error: can't upload dark/stdlib/Test::noop_v0 (too low, 0 <
[something])
[x] error: can't re-upload dark/stdlib/Test::noop_v[something], it's
already in the db

Doesn't test:
- package_fns in the db show up in autocomplete
- can run in repl
- can run in the bwd context
For all of these, I'd want to be able to switch canvases mid-test (to
upload the fn in one canvas and use it in another). But it might be
simpler to just add a package_v0 record to the db and test that way, in
a different set of integration tests

NOTE: we now also have the ability to use the test_admin user in tests,
if the test name contains the string '_as_admin'. This was necessary
since upload is currently limited to Dark employees.

Remove comment we're not going to TODO