This fixes the CSS OM/Experimental Web Platform features
The CSS Object Model View Module spec proposes to change MouseEvent
coordinates to be double instead of int (long). Chrome's "Experimental
Web Plaform features" flag causes this to be active on OS X. (We believe
- ismith was unable to repro on Linux, Julian and Victoria both saw it
on OS X.)

Problem is, our decoders assume these coordinates are ints, and so mouse
clicks in browsers with the right config result in "undecodable json"
errors in the dev console, and clicks appear (to users) to be ignored.

Since this is in chrome://flags, it's a bit of an edge case for now, but
we've received 2-3 customer reports. And presumably eventually this will
be a more standard spec and configuration, so we want to fix it now.

We decode these numerics in `ViewUtil.ml`'s `decodeClickEvent`, so I
started by replacing the int decoder with a float decoder. I then just
updated types as the compiler indicated.

(This does not break browsers that use int coords, b/c js/json does not
distinguish numerics that are ints vs floats, and so an "int" input can
be decoded into a float.)

There are a few places where we call `float_of_int` because Tea.Mouse
gives us int values, or `truncate`, because we have a fn that wants ints
(`Random.range` for ~random placement of TLs is an example of this).
I think these should be fine, but I recognize that this is slightly
inconsistent.

https://trello.com/c/fjFvU8JY/2328-fix-the-side-effects-of-upcoming-changes-to-the-css-object-model-which-passes-mouse-positions-as-doubles?menu=filter&filter=member:iansmith286