Post-auth0, we should not be putting password hashes in postgres
We need to keep the field in the DB for the local env to work, but
DarkInternal:: functions for inserting and upserting users should use
Password.invalid.

I know we normally bump function versions; I opted not to do so here,
because the type sig is unchanged, and we have (or should have) no code
still using the string password returned by these functions.

(If we _do_, those passwords are invalid for login anyway, so this does
not break anything.)

Post-deploy, I'll `UPDATE accounts SET password = ''` in production.

https://trello.com/c/OJ7kiPF0/2348-post-rollout-darkinternalinsertuser-and-upsertuser-should-not-set-passwords-and-all-prod-password-hashes-should-be-empty