flags: split ast instead of token streams
In several cases, we need the actual AST (FluidExpression.t) of the
panel in order to properly compute this or that about the caret
placement or edits and things. This is impossible if the panels only
have their token streams (tokenInfo list) and not the actual ast.

So, now we pluck out the feature flags from the AST and tokenize each
subtree just-in-time.

This temporarily breaks a bunch of caret placement stuff, which again has no notion
of which panel it's operating against.