fix Vdom warnings and weird rendering of sidebar
Because the sidebar uses a bunch of SVGs, having lots of nodes be
reconstructed by BSTea when it's collapsed/expanded causes 1) Vdom
warnings and 2) weird rendering as the SVGs are loaded back from
cache — sometimes it takes a few frames for them to show up.

Instead, just always render both sidebars and toggle the visibility,
which fixes both problems.