do wrapper shenanigans with feature flag panel too
all the same math we do to figure out the correct caret / selection
position for the wrapper has to be done for the FF panel too. We can't
just naively add the offset and hope it works.

This allows us to automatically run editing tests in a FF panel as well,
which works for all but 2 tests!