We used two versions of allCompletions in fluid, reduce to 1
There was fluidAutocompleteState.allCompletions (record field) and
FluidAutocomplete.allCompletetions (module function). They basically did the
same thing but one was stateful. Removing one broke all the tests so I removed
the other and that was fine.

Basically, we used the record field to sometimes hold state that wasn't always
valid, and its main function was to temporarily store the list when we passed
the autocomplete into refilter.