Canvas cloning code + script
The end goal is that, on account creation, we clone canvas
sample-getting-started into the new user's account
(/a/ismith-getting-started). In addition, the source canvas may be
updated - extant clones will not update, but subsequent clones/accounts
will use the latest version of sample-getting-started.

Secondary features:
- we edit the ops before writing them - undo history is not valuable
  here (esp since we have no git-analogue for examining history), and
  removing it de-risks "I put something in that was sensitive or incorrect
  and then removed it"
- we also walk the ops' ASTs and check for string literals referring to
  the from_canvas and update them. So if you made an Http call to
  sample-getting-started.builtwithdark.com, the new canvas will call
  <username>-getting-started.builtwithdark.com

Of note:
- we check when we start that from_canvas exists ("cloning from an empty
  canvas" makes no sense) and to_canvas does not ("don't clobber an
  existing canvas")
- we do all inserts in a transaction so canvas creation + op saving are
  atomic
- backend/bin/clone_canvas.ml is a script using this same code to clone
  between arbitrary canvases; primarily useful for development, but I
  could imagine a use case like "user wants their canvas moved into (or
  out of) an org", or "user wants a new username, easiest if we just clone
  their canvases over to the new username"

Discussed, but determined was an acceptable flaw:
If Ellen is editing the source canvas, any accounts created during that
period will get the current state - which may be mid-edit! We discussed
in Slack, there's really no good options in Dark for setting a
point-in-time to clone from, so we're calling this ok for now. Ellen
notes that she does not expect to edit sample-getting-started
frequently, and is willing to both (a) attempt to do it during
low-traffic periods, and (b) accept that some small number of users may
get a "WIP" getting-started canvas.

Spec: https://www.notion.so/darklang/Getting-Started-Canvas-3b96699611594ef887fcc1f4aab8e4a0
Trello: https://trello.com/c/k1nMID2e/2597-getting-started-canvas