Add FluidMouseDoubleClick
The initial goal was to do specific behaviour on doubleclick (if you
doubleclicked on a function, you'd select the entire function). However, the
click and doubleclick handlers were hopelessly intertwined, so step 1 was to
prise them apart.

Then I was able to use getFluidSelectionRange for the doubleclick handler, which
improved the selection considerably (for example, in the past, if you
doubleclicked a function call after the ::, it only selected the version number;
now it selections the function name and version number (though not the module name))