Make custom-domain routing happen in the DB
This is part of the custom-domain project.

Today (status quo, pre-this-PR), adding a custom domain requires
modifying webserver.ml and doing a deploy of the app.

Instead, we'd like to have a database table, custom_domains, that has a
host ("dark-001.metaforgotten.org") and a canvas
("ismith-customdomain-001") and routes accordingly. In phase 1 of the
custom-domains project, editing this table is done by dark employees by
script - the goal is not yet self-service. (Said script is not included
in this PR - I've tested locally using pgcli, but this will be wrapped
in the new add-custom-domain script being worked on in the context of
the new TLS branch.)

`host` must be unique and indexed; it is our primary key. `canvas` may
be a dupe - you could have multiple hostnames pointing at the same
canvas.

We're using canvas:text because that's how webserver.ml does routing (by
name); there are a number of places we use canvas_name as an fk anyway.
Additionally, the CNAME target depends on the canvas name. For now, all
of this is fine because we can't change canvas' names anyway.

For reasons of backwards compatibility, we're going to leave in place
any hardcoded values (eg for api.venu.fm), and we'll use those before
falling back to the database. This also allows us to deploy this PR in
production now, without blocking on other custom-domain cert work. That
is desirable because that work is happening in a second cluster
("sandbox"); we have not created a new database for that sandbox
cluster, and by deploying this db migration to prod, we won't have to.

https://trello.com/c/W3bQhXBm/2768-custom-domains-implementation