Address the BadConfig case
If you manage to double-patch the darkcustomdomain-l4-ingress, you will
see this in the event log:
```
 Warning  BadConfig          9s                 cert-manager              Duplicate TLS entry for secretName "dark-002.metaforgotten.org-tls"
```

This does not break existing certs, but it does prevent further certs
from being created. I've added some checks to reduce the odds of this
happening, and also documented steps for fixing if we do end up in that
situation.