Remove autocomplete update from the start of Fluid.updateMsg
This takes about 1.5ms (5-10% of the cost of keypress), and happens on every
keystroke.

This was only needed for tests, so I moved it to the test framework.

I validated that this continues to work when switching between handlers, which
was the only place I imaged could suffer from this.