Run tunnel in localhost env
It's "just" an ssh tunnel from localhost to localhost, but libcurl's
logs (seen in backend/libbackend/curl_logging.ml) look different when we
use a socks5 proxy (as we do in prod) vs when we don't (as in local,
prior to this commit).

This led to a bug in https://trello.com/c/ov1NRfTp/2963-add-ip-port-to-libcurl-logs, where we logged primary_ip not realizing that it'd always be the ip of the socks5 proxy, our tunnel service.

https://trello.com/c/mnqpYtSo/2964-consider-using-the-tunnel-socks5-proxy-service-in-dev-ci-envs-for-consistency-across-environments