Disallow hitting 169.254.* from the tunnel
This means that http://metadata will no longer be reachable in Darklang
by using `HttpClient::get`.

Tested:
- deployed a new cluster
- confirmed, pre-this-commit, that I could `kubectl exec` into the
  tunnel and curl both http://metadata and http://neverssl.com
- deployed the network policy in this PR
- confirmed I could not curl http://metadata anymore, but
  http://neverssl.com still works

No special action needed to get the NetworkPolicy picked up - I did not
have to, eg, restart the tunnel pod. Nor did I have to enable
NetworkPolicies in cloud console web ui; that was mentioned in the
original NetworkPolicy PRs (see #72), and is/was a cluster-level
setting, but did not seem to be necessary when I ran gke-create-cluster;
I assume we copied it over from the prod cluster or something.

Post-deploy, I'll confirm that HttpClient::get works for known-good
cases (google.com, neverssl.com) and fails for http://metadata. That
closes the hole.