Data analysis scripts for Victoria
Right now, there is data work that Victoria (and Ellen) want to do that
they have to ask engineering to do because we lack tooling, and we're
not currently in a position to build out some of that analytics tooling.

Often this results in me doing one-off work, which may or may not be a
script; this PR is the beginning of having a place to preserve some of
those scripts, enabling Victoria to self-service some of these requests
for data. (And if not, at least letting engineering not reinvent the
wheel.)

First script in this folder:

During last week's security incident, Ellen and Victoria came up with a
set of usernames and canvases. and wanted email addresses to go with
each. I wrote a script to take the spreadsheet they provided and enhance
it wiih email addresses from prodclone. Putting it here is an attempt to
preserve it for future use.

It's in Ruby because that's the language I am most comfortable scripting
in under time pressure; I did take steps to avoid using non-stdlib
libraries, and it runs in Ruby 2, which is default on OS X these days.

To come later:
Training /w victoria, which may or may not result in written docs. We'll
see what's necessary and pragmatic.

https://trello.com/c/oUECRrfM/3035-preserve-the-data-analysis-script-from-security-incident-for-victorias-use