Cookie and authorization headers are sensitive, don't log to honeycomb in raw form
We could just use honeytail's --drop_field flag, but let's instead use
--scrub_field, which replaces them with a SHA256 hash.

The ticket said "ideally we should whitelist the headers that go to
HC", which I agree with, and we do (in scripts/support/nginx.conf, we
set a log_format which must match the format in
scripts/support/kubernetes/builtwithdark/honeycomb.yaml).

(Why scrub vs removing from both log_format lines? B/c it may be useful
to be able to look for specific sessions, or count sessions.)

https://trello.com/c/Hy4HrdVM/3042-authorization-and-cookie-headers-should-not-go-to-honeycomb-in-raw-form