Remove maxim's special qws (which no longer exist) from tunnel ingress
The tunnel service's network policy only allows ingress from pods with
particular labels - that is, only our ocaml apps. (bwd-app, qw-worker,
cron-checker, editor-app.)

Shortly before strangeloop (2019), we had problems where a particular user was
overwhelming the queue workers and consuming all the compute. As a
short-term fix, we launched a special deployment of qw pods just for
that user, and so we had to add that pod to the allowed ingresses.

We eventually fixed the problem and took down the
special-one-user-deployment, but the tunnel network policy was
overlooked.

This change is a no-op; it's just removing obsolete cruft.